// projects/rivet-style/src/lib/rivet-services/errors/error-types.ts
export enum ErrorTypes {
  Email = 'email',
  CannotHaveSpaces = 'cannotHaveSpaces',
  OnlySpaces = 'onlySpaces',
  NameTaken = 'nameTaken',
  SpecialCharacters = 'specialCharacters',
  RequireAutoCompleteMatch = 'requireAutoCompleteMatch',
  ValidateNumber = 'validateNumber',
  IsValidMaterialDatepickerDate = 'matDatepickerParse',
  MaterialDatepickerMax = 'matDatepickerMax',
  MaterialDatepickerMin = 'matDatepickerMin',
  Required = 'required'
}